requirements.txt
pathspec
GitPython
